JAVA-41593 core-java-io-5 module Splitted and article moved to core-java-io-2 module
