+++ core-java-modules/core-java-io-2/README.md
@@ -11,4 +11,5 @@ This module contains articles about core Java input and output (IO)
 - [Create a Directory in Java](https://www.baeldung.com/java-create-directory) 
 - [Java IO vs NIO](https://www.baeldung.com/java-io-vs-nio)
 - [Creating Temporary Directories in Java](https://www.baeldung.com/java-temp-directories)
+- [Convert an OutputStream to a Byte Array in Java](https://www.baeldung.com/java-outputstream-byte-array)
 - [[<-- Prev]](/core-java-modules/core-java-io)[[More -->]](/core-java-modules/core-java-io-3)
