+++ core-java-modules/core-java-io-5/README.md
@@ -9,10 +9,8 @@ This module contains articles about core Java input and output (IO)
 - [How to Write Strings to OutputStream in Java](https://www.baeldung.com/java-write-string-outputstream)
 - [Read a File and Split It Into Multiple Files in Java](https://www.baeldung.com/java-read-file-split-into-several)
 - [Read and Write Files in Java Using Separate Threads](https://www.baeldung.com/java-read-write-files-different-threads)
-- [Convert an OutputStream to a Byte Array in Java](https://www.baeldung.com/java-outputstream-byte-array)
 - [Reading a .gz File Line by Line Using GZIPInputStream](https://www.baeldung.com/java-gzipinputstream-read-gz-file-line-by-line)
 - [Opening HTML File Using Java](https://www.baeldung.com/java-open-html-file)
 - [PrintWriter write() vs print() Method in Java](https://www.baeldung.com/java-printwriter-write-vs-print)
 - [Format Output in a Table Format Using System.out](https://www.baeldung.com/java-format-output-table-system-out)
 - [[<-- Prev]](/core-java-modules/core-java-io-4)
-
